Squashed commit of the following:

commit cfe8babcf3ae41c7f80f43a24aa96b41c52e68fc
Author: StasGorobzeev <stanislav.gorobzeev@upholding.biz>
Date:   Mon Sep 2 13:13:37 2019 +0300

    fix

+++ Model/src/main/java/me/exrates/model/dto/merchants/adgroup/AdGroupRequestRefillBodyDto.java
@@ -12,7 +12,7 @@ public class AdGroupRequestRefillBodyDto {
 
     private BigDecimal amount;
 
-    private Integer tel;
+    private Long tel;
 
     private String platform;
